requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2
kafka-python==2.0.2
prometheus-client==0.19.0
python-json-logger==2.0.7
pymongo==4.6.1
aiohttp==3.9.1
